Prevent local variables from shadowing of implicit member references
E.g. `Entity#posX` is remapped to `x` which may then be shadowed by
a local variable with the same name.

Since I cannot figure out how to get JDT to give me a list of all
variables in scope at a particular node, the solution employed is to
just forbid any implicit references to remapped members.
This finds quite a few false positives (but not too many to manually
deal with), so I'd be quite happy to switch to another solution if
one becomes available.