Fix fully qualified constructor call in Kotlin
Prior to this commit the `a.pkg.A()` in the
`remaps fully qualified class references` kotlin test case would be
remapped as `a.pkg.B()` instead of `b.pkg.B()`.
Everything else and all other test cases were passing even before this
commit; they were added as part of this commit to prevent regression and
increase test coverage.