Fix remapping of Shadow/Invoker methods with mappings only in target class super class/interface
E.g. in the test cases, `AMixin` contains a `Shadow`/`Invoker` method
targeting `aInterfaceMethod`. That method does exist in the `A` class
which the mixin targets, but its mapping to `bInterfaceMethod` only
exists in the `AInterface` class which it is declared in.
Previously we only looked up mappings directly in `A` so we wouldn't see
this mapping and fail to remap these Shadow/Invoker methods. This commit
fixes that by checking the mappings of all parent classes/interfaces of
the mixin target class as well.